% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_antismash.R
\name{get_antismash_features}
\alias{get_antismash_features}
\title{Extract features from Antismash json files}
\usage{
get_antismash_features(jsons, features_types = NULL)
}
\arguments{
\item{jsons}{A character vector of valid filepaths.}

\item{features_types}{A list of desired features types, by default all.}
}
\value{
A list of list of records.
}
\description{
Extract features from Antismash json files
}
