% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_antismash.R
\name{get_antismash_regions}
\alias{get_antismash_regions}
\title{Creates a dataframe with region information from antiSMASH data}
\usage{
get_antismash_regions(features)
}
\arguments{
\item{features}{List object generated by \code{get_antismash_features}}
}
\value{
A tibble where each where each region is an observation
}
\description{
Creates a dataframe with region information from antiSMASH data
}
