% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_antismash.R
\name{get_bigscape_clustering}
\alias{get_bigscape_clustering}
\title{Extract clustering information from bigscape}
\usage{
get_bigscape_clustering(bigscape_dir)
}
\arguments{
\item{bigscape_dir}{A valid filepath}
}
\value{
A DataFrame
}
\description{
Extract clustering information from bigscape
}
