% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_antismash.R
\name{get_bigscape_networks}
\alias{get_bigscape_networks}
\title{Extract network information from bigscape}
\usage{
get_bigscape_networks(bigscape_dir, cutoff = 0.3)
}
\arguments{
\item{bigscape_dir}{A valid filepath}

\item{cutoff}{A numeric value indicating the cutoff.}
}
\value{
A DataFrame
}
\description{
Extract network information from bigscape
}
