% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wants_hue.R
\name{get_wants_hue}
\alias{get_wants_hue}
\title{Generate a palette using I Want Hue}
\usage{
get_wants_hue(n, seed = 1, palette = "colorblind_friendly")
}
\arguments{
\item{n}{A numeric vector of one with the desired number of colors}

\item{seed}{A numeric vector of one with the desired seed (palettes are generated randomly)}

\item{palette}{A character vector of one with name of the palette you want to use.
To see all available palettes run \code{names(rwantshue::hcl_presets)}.}
}
\value{
A character vector with the hex codes.
}
\description{
This function allows you to generate a palette of colors using the web
https://medialab.github.io/iwanthue/
}
\examples{
get_wants_hue(n = 5)
get_wants_hue(n = 5, seed = 2)
get_wants_hue(n = 5, seed = 2, palette = "fluo")
}
