% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_physeq_tax_label_unique.R
\name{make_physeq_tax_label_unique}
\alias{make_physeq_tax_label_unique}
\title{Make unique taxonomic labels from a Phyloseq taxonomyTable}
\usage{
make_physeq_tax_label_unique(taxtable)
}
\arguments{
\item{taxtable}{A phyloseq-class instance}
}
\value{
A character vector with unique taxonomic labels.
}
\description{
This function will extract the available deepest taxonomic label
from a phyloseq taxonomyTable. You still need to assign that
(see example below).
}
\examples{
taxtable <- structure(
  c("Bacteria", "Proteobacteria", "Deltaproteobacteria","", "", "", ""),
  dim = c(1L, 7L),
  dimnames = list(
  "sp1", c("Kingdom", "Phylum", "Class","Order",
   "Family", "Genus", "Species")
   ))
make_physeq_tax_label_unique(taxtable)
}
