% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_template_toy_data.R
\docType{data}
\name{phyloseq_template_toy_data}
\alias{phyloseq_template_toy_data}
\title{Phyloseq template toy data}
\format{
\subsection{\code{phyloseq_template_toy_data}}{

A list containing 3 DataFrames:
\describe{
\item{meta}{Metadata table with 29 samples and 4 features (layer, age, pH and depth).}
\item{otu}{OTU table with 29 samples and 3,728 taxa groups.}
\item{taxa}{Taxonomic table with 7 ranks and 3,728 taxa groups.}
...
}
}
}
\usage{
phyloseq_template_toy_data
}
\description{
A subset of data to illustrate how to perform an analysis of
OTU with phyloseq, vegan and microbiome.
}
\keyword{datasets}
