% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_silva_prefix_from_taxtable.R
\name{remove_silva_prefix_from_taxtable}
\alias{remove_silva_prefix_from_taxtable}
\title{Removes Silva prefix from a Phyloseq taxonomyTable}
\usage{
remove_silva_prefix_from_taxtable(taxtable)
}
\arguments{
\item{taxtable}{A taxonomy table from phyloseq (obtained using \code{tax_table()})}
}
\value{
A character vector with unique taxonomic labels.
}
\description{
This function will remove the silva prefix from a Phyloseq taxonomy table.
You still need to assign that
}
\examples{
\dontrun{
taxtable <- structure(c("k__Bacteria", " p__Proteobacteria",
"c__Deltaproteobacteria", " o__Oligoflexales", " f__0319-6G20",
"g__", " s__"), dim = c(1L, 7L),
dimnames = list("sp1", c("Kingdom",
"Phylum", "Class", "Order", "Family", "Genus", "Species")))
remove_silva_prefix_from_taxtable(taxtable)
}
}
