% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_AU_ENVS.R
\name{set_theme_AU_ENVS}
\alias{set_theme_AU_ENVS}
\title{Set default AU-ENVS R theme
This function will override some layout options that will
affect how the graphics look. It has side effects.}
\usage{
set_theme_AU_ENVS(locale = "English_Denmark")
}
\arguments{
\item{locale}{A character vector of one indicating the desired locale.
Most of the time, the default "English_Denmark" is desired, although
sometimes you may want to use the "Denmark_English" option for
Danish months,  names, etc...}
}
\description{
Set default AU-ENVS R theme
This function will override some layout options that will
affect how the graphics look. It has side effects.
}
\examples{
\dontrun{set_theme_AU_ENVS()}
\dontrun{set_theme_AU_ENVS("English_Denmark")}
}
